Put navigation links data in a shared place
