+++ app/data.js
@@ -1,0 +1,36 @@
+import { image } from "./filters";
+
+export var NAV_LINKS = [
+  {
+    name: "Chat Room",
+    url: "https://gitter.im/pytexas/PyTexas",
+    external: true,
+    icon: image("img/icons/chat.svg")
+  },
+  {
+    name: "Sponsor",
+    url: "/page/sponsors/prospectus",
+    icon: image("img/icons/sponsors.svg")
+  },
+  {
+    name: "About",
+    url: "/page/about/conference",
+    icon: image("img/icons/about.svg")
+  },
+  {
+    name: "Venue",
+    url: "/page/venue/map",
+    icon: image("img/icons/venue.svg")
+  },
+  {
+    name: "Community",
+    url: "/page/community/meetups",
+    icon: image("img/icons/community.svg")
+  },
+  {
+    name: "Blog",
+    url: "https://medium.com/pytexas",
+    external: true,
+    icon: image("img/icons/blog.svg")
+  }
+];
